deduplicated compaction tests
