+++ test/src/main/java/org/apache/accumulo/test/compaction/ExternalCompaction_2_IT.java
@@ -25,6 +25,7 @@ import static org.apache.accumulo.test.compaction.ExternalCompactionTestUtils.MA
 import static org.apache.accumulo.test.compaction.ExternalCompactionTestUtils.compact;
 import static org.apache.accumulo.test.compaction.ExternalCompactionTestUtils.confirmCompactionCompleted;
 import static org.apache.accumulo.test.compaction.ExternalCompactionTestUtils.confirmCompactionRunning;
+import static org.apache.accumulo.test.compaction.ExternalCompactionTestUtils.countTablets;
 import static org.apache.accumulo.test.compaction.ExternalCompactionTestUtils.createTable;
 import static org.apache.accumulo.test.compaction.ExternalCompactionTestUtils.row;
 import static org.apache.accumulo.test.compaction.ExternalCompactionTestUtils.waitForCompactionStartAndReturnEcids;
@@ -49,7 +50,9 @@ import org.apache.accumulo.core.client.TableNotFoundException;
 import org.apache.accumulo.core.compaction.thrift.TCompactionState;
 import org.apache.accumulo.core.conf.Property;
 import org.apache.accumulo.core.data.TableId;
+import org.apache.accumulo.core.metadata.MetadataTable;
 import org.apache.accumulo.core.metadata.schema.ExternalCompactionId;
+import org.apache.accumulo.core.metadata.schema.MetadataSchema;
 import org.apache.accumulo.core.metadata.schema.TabletMetadata.ColumnType;
 import org.apache.accumulo.core.metadata.schema.TabletsMetadata;
 import org.apache.accumulo.harness.MiniClusterConfigurationCallback;
@@ -149,15 +152,28 @@ public class ExternalCompaction_2_IT extends SharedMiniClusterBase {
           .confirmCompactionRunning(getCluster().getServerContext(), ecids);
       assertTrue(matches > 0);
 
+      // when the compaction starts it will create a selected files column in the tablet, wait for
+      // that to happen
+      while (countTablets(getCluster().getServerContext(), table1,
+          tm -> tm.getSelectedFiles() != null) == 0) {
+        Thread.sleep(1000);
+      }
+
       client.tableOperations().cancelCompaction(table1);
 
       confirmCompactionCompleted(getCluster().getServerContext(), ecids,
           TCompactionState.CANCELLED);
 
-      // We need to cancel the compaction or delete the table here because we initiate a user
-      // compaction above in the test. Even though the external compaction was cancelled
-      // because we split the table, FaTE will continue to queue up a compaction
-      client.tableOperations().cancelCompaction(table1);
+      // ensure the canceled compaction deletes any tablet metadata related to the compaction
+      while (countTablets(getCluster().getServerContext(), table1,
+          tm -> tm.getSelectedFiles() != null || !tm.getCompacted().isEmpty()) > 0) {
+        Thread.sleep(1000);
+      }
+      //
+      // // We need to cancel the compaction or delete the table here because we initiate a user
+      // // compaction above in the test. Even though the external compaction was cancelled
+      // // because we split the table, FaTE will continue to queue up a compaction
+      // client.tableOperations().cancelCompaction(table1);
     }
   }
 
@@ -181,10 +197,25 @@ public class ExternalCompaction_2_IT extends SharedMiniClusterBase {
       int matches = confirmCompactionRunning(getCluster().getServerContext(), ecids);
       assertTrue(matches > 0);
 
+      // when the compaction starts it will create a selected files column in the tablet, wait for
+      // that to happen
+      while (countTablets(getCluster().getServerContext(), table1,
+          tm -> tm.getSelectedFiles() != null) == 0) {
+        Thread.sleep(1000);
+      }
+
       client.tableOperations().delete(table1);
 
       confirmCompactionCompleted(getCluster().getServerContext(), ecids,
           TCompactionState.CANCELLED);
+
+      // ELASTICITY_TODO make delete table fate op get operation ids before deleting
+      // there should be no metadata for the table, check to see if the compaction wrote anything
+      // after table delete
+      try (var scanner = client.createScanner(MetadataTable.NAME)) {
+        scanner.setRange(MetadataSchema.TabletsSection.getRange(tid));
+        assertEquals(0, scanner.stream().count());
+      }
 
     }
   }
